Moved every component(Profile, My activity, Settings) to each folder and file
